requirements.txt
# Data & Analysis
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3

# Technical Indicators
ta>=0.10.2
pandas-ta>=0.3.14b

# Backtesting
vectorbt>=0.26.0

# Notifications
line-bot-sdk>=3.0.0
yagmail>=0.15.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0
mplfinance>=0.12.10

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Development
pytest>=7.0.0
